+++ R/emmGrid-methods.R
@@ -205,7 +205,7 @@ vcov.emmGrid = function(object, ...) {
 #' and confidence intervals.}
 #' 
 #' \item{\code{famSize}}{(integer) is the number of means involved in a family of
-#' inferences; used in Tukey and Scheffe adjustments}
+#' inferences; used in Tukey adjustment}
 #' 
 #' \item{\code{infer}}{(\code{logical} vector of length 2) is the default value
 #' of \code{infer} in \code{\link{summary.emmGrid}}.}
@@ -222,6 +222,9 @@ vcov.emmGrid = function(object, ...) {
 #' 
 #' \item{\code{side}}{(numeric or character) \code{side} specification for for
 #' \code{summary} or \code{test} (taken to be zero if missing).}
+#' 
+#' \item{\code{sigma}}{(numeric) Error SD to use in predictions and for bias-adjusted
+#' back-transformations}
 #' 
 #' \item{\code{delta}}{(numeric) \code{delta} specification for \code{summary}
 #' or \code{test} (taken to be zero if missing).}
@@ -285,7 +288,7 @@ update.emmGrid = function(object, ..., silent = FALSE) {
     valid.misc = c("adjust","alpha","avgd.over","by.vars","delta","df",
                    "initMesg","estName","estType","famSize","infer","inv.lbl",
                    "level","methDesc","nesting","null","predict.type","pri.vars"
-                   ,"side","tran","tran.mult","tran2","type","is.new.rg")
+                   ,"side","sigma","tran","tran.mult","tran2","type","is.new.rg")
     valid.slots = slotNames(object)
     valid.choices = union(valid.misc, valid.slots)
     misc = object@misc
@@ -391,6 +394,12 @@ update.emmGrid = function(object, ..., silent = FALSE) {
 #'   displayed is just enough to reasonably distinguish estimates from the ends
 #'   of their confidence intervals; but always at least 3 digits. If
 #'   \code{FALSE}, the system value \code{getOption("digits")} is used.}
+#' \item{\code{back.bias.adj}}{A logical value controlling whether we 
+#'   try to adjust bias when back-transforming. If \code{FALSE}, we use naive
+#'   back transformation. If \code{TRUE} \emph{and \code{sigma} is available}, a
+#'   second-order adjustment is applied to estimate the mean on the response
+#'   scale.}
+#'   
 #' }%%% end describe{}
 #' Some other options have more specific purposes:
 #' \describe{
@@ -477,6 +486,7 @@ emm_defaults = list (
     msg.nesting = TRUE,       # message when nesting is detected
     estble.tol = 1e-8,        # tolerance for estimability checks
     simplify.names = TRUE,    # simplify names like data$x to just "x"
+    back.bias.adj = FALSE,    # Try to bias-adjust back-transformations?
     opt.digits = TRUE,        # optimize displayed digits?
     lmer.df = "kenward-roger",  # Use Kenward-Roger for df
     disable.pbkrtest = FALSE, # whether to bypass pbkrtest routines for lmerMod
@@ -532,8 +542,18 @@ emm_defaults = list (
 #'   "log"}, and is used to label the predictions if subsequently summarized
 #'   with \code{type = "response"}.
 #' @param predict.type Character value. If provided, the returned object is
-#'   updated with the given type, e.g., \code{"response"}. 
-#'   See \code{\link{update.emmGrid}}.
+#'   updated with the given type to use by default by \code{summary.emmGrid}
+#'   (see \code{\link{update.emmGrid}}).  This may be useful if, for example,
+#'   when one specifies \code{transform = "log"} but desires summaries to be
+#'   produced by default on the response scale.
+#' @param bias.adjust Logical value for whether to adjust for bias in
+#'   back-transforming (\code{transform = "response"}). This requires a value of 
+#'   \code{sigma} to exist in the object or be specified.
+#' @param sigma Error SD assumed for bias correction (when 
+#'   \code{transform = "response"} and a transformation
+#'   is in effect). If not specified,
+#'   \code{object@misc$sigma} is used, and an error is thrown if it is not found.
+#' @param ... Ignored.
 #'   
 #' @note Another way to use \code{regrid} is to supply a \code{transform} 
 #'   argument to \code{\link{ref_grid}} (either directly of indirectly via
@@ -556,7 +576,8 @@ emm_defaults = list (
 #' emmeans(pigs.lm, "source", transform = "response")
 #' # NOTE: pairs() of the above will be DIFFERENCES of these results
 regrid = function(object, transform = c("response", "mu", "unlink", "log", "none"), 
-                  inv.log.lbl = "response", predict.type) 
+                  inv.log.lbl = "response", predict.type, 
+                  bias.adjust = get_emm_option("back.bias.adj"), sigma, ...) 
 {
     if (is.logical(transform))   # for backward-compatibility
         transform = ifelse(transform, "response", "none")
@@ -602,9 +623,18 @@ regrid = function(object, transform = c("response", "mu", "unlink", "log", "none
     }
     
     if(transform %in% c("response", "mu", "unlink", "log") && !is.null(object@misc$tran)) {
-        link = attr(est, "link")
-        D = .diag(link$mu.eta(object@bhat[estble]))
-        object@bhat = link$linkinv(object@bhat)
+        flink = link = attr(est, "link")
+        if (bias.adjust) {
+            if(missing(sigma))
+                sigma = object@misc$sigma
+            link = .make.bias.adj.link(link, sigma)
+            if (!is.na(PB[1])) # special frequentist version when sigma is MCMC sample
+                flink = .make.bias.adj.link(flink, mean(sigma))
+            else
+                flink = link
+        }
+        D = .diag(flink$mu.eta(object@bhat[estble]))
+        object@bhat = flink$linkinv(object@bhat)
         object@V = D %*% tcrossprod(object@V, D)
         if (!is.na(PB[1]))
             PB = matrix(link$linkinv(PB), ncol = ncol(PB))
@@ -621,6 +651,7 @@ regrid = function(object, transform = c("response", "mu", "unlink", "log", "none
         }
         else
             object@misc$tran = object@misc$tran.mult = object@misc$inv.lbl = NULL
+        sigma = object@misc$sigma = NULL
     }
     if (transform == "log") { # from prev block, we now have stuff on response scale
         Vee = vcov(object)