+++ R/emmeans-package.R
@@ -34,8 +34,7 @@
 #' @section Overview:
 #' \describe{
 #' \item{Vignettes}{A number of vignettes are provided to help the user get
-#' acquainted with the \pkg{emmeans} package and see some examples. 
-#' See the \href{../doc/index.html}{vignette index}.}
+#' acquainted with the \pkg{emmeans} package and see some examples.}
 #' 
 #' \item{Concept}{Estimated marginal means (see Searle \emph{et al.} 1980 are
 #' popular for summarizing linear models that include factors. For balanced