+++ R/interfacing.R
@@ -37,7 +37,7 @@
 #' functions of the regression coefficients for a given grid of predictor
 #' values. These tasks are performed by calls to \code{recover_data} and
 #' \code{emm_basis} respectively. A vignette giving details and examples
-#' is available via \href{../doc/extending.pdf}{vignette("extending", "emmeans")}
+#' is available via \href{../doc/xtending.html}{vignette("xtending", "emmeans")}
 #' 
 #' To extend \pkg{emmeans}'s support to additional model types, one need only
 #' write S3 methods for these two functions. The existing methods serve as
@@ -84,7 +84,7 @@
 #'    In those cases, users should be careful to provide the actual data
 #'    used to fit the model in the \code{data} argument.
 #'   
-#' @seealso \href{../doc/extending.pdf}{Vignette on extending emmeans}
+#' @seealso \href{../doc/xtending.html}{Vignette on extending emmeans}
 #' 
 #' @export
 recover_data = function(object, ...) {
@@ -234,6 +234,12 @@ recover_data.call = function(object, trms, na.action, data = NULL, params = NULL
 #'   \code{dffun}}.
 #' } %%% end of describe
 #' @export
+#' 
+#' @section Communication between methods:
+#' If the \code{recover_data} method generates information needed by \code{emm_basis},
+#' that information may be incorporated by creating a \code{"misc"} attribute in the
+#' returned recovered data. That information is then passed as the \code{misc} 
+#' argument when \code{ref_grid} calls \code{emm_basis}.
 #' 
 #' @section Optional hooks:
 #' Some models may need something other than standard linear estimates and